+++ tests/lexer_test.cpp
@@ -3,11 +3,12 @@
 #include <vector>
 #include "lexer/tokenization.hpp"
 #include "lexer/tokens.hpp"
+#include "types.hpp"
 
 TEST(lexerTests, HandlesOperators) {
     std::string op = "+";
-    uint64_t expected_line = 1;
-    uint64_t expected_column = 1;
+    u16 expected_line = 1;
+    u16 expected_column = 1;
     std::vector<Token> expectedOutput = {
         Token{TokenType::Plus, op, expected_line, expected_column}};
